@@ -1,4 +1,4 @@
-package com.zylex.streamapi;
+package ru.panyukovnn.streamapi.streamapi;
 
 import org.junit.Test;
 
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.*;
 
-import static com.zylex.streamapi.Util.printStreamInLine;
+import static ru.panyukovnn.streamapi.streamapi.Util.printStreamInLine;
 
 public class Creation {
 
@@ -32,6 +32,7 @@ public class Creation {
         Map<Integer, String> map = Map.of(0, "a", 1, "b");
 
         Stream<Map.Entry<Integer, String>> mapStream = map.entrySet().stream();
+        Stream<Integer> mapKeyStream = map.keySet().stream();
     }
 
     @Test
@@ -50,15 +51,15 @@ public class Creation {
     public void parallel() {
         List<Integer> list = List.of(100, 200, 300, 400);
         Stream<Integer> integerStream1 = list.parallelStream()
-                .filter(x -> x > 100)
-                .map(x -> x * 2); // 400 600 800
+            .filter(x -> x > 100)
+            .map(x -> x * 2); // 400 600 800
 
         printStreamInLine(integerStream1);
 
         Stream<Integer> integerStream2 = list.stream()
-                .parallel()
-                .filter(x -> x > 200)
-                .map(x -> x * 10); // 3000 4000
+            .parallel()
+            .filter(x -> x > 200)
+            .map(x -> x * 10); // 3000 4000
 
         printStreamInLine(integerStream2);
     }
@@ -87,7 +88,7 @@ public class Creation {
     @Test
     public void generate() {
         Stream<Integer> limit = Stream.generate(() -> 6)
-                .limit(6);
+            .limit(6);
 
         printStreamInLine(limit); // 6, 6, 6, 6, 6, 6
     }
@@ -95,7 +96,7 @@ public class Creation {
     @Test
     public void iterate() {
         Stream<Integer> iterate = Stream.iterate(2, x -> x + 6)
-                .limit(6);
+            .limit(6);
 
         printStreamInLine(iterate); // 2, 8, 14, 20, 26, 32
     }
@@ -110,17 +111,26 @@ public class Creation {
     @Test
     public void concat() {
         Stream<Integer> concat = Stream.concat(
-                Stream.of(1, 2, 3),
-                Stream.of(4, 5, 6));
+            Stream.of(1, 2, 3),
+            Stream.of(4, 5, 6));
 
         printStreamInLine(concat); // 1, 2, 3, 4, 5, 6
+
+        List<Integer> list1 = List.of(1, 2, 3);
+        List<Integer> list2 = List.of(4, 5, 6);
+        Stream<Integer> concat2 = Stream.concat(
+            list1.stream(),
+            list2.stream()
+        );
+
+        printStreamInLine(concat2);
     }
 
     @Test
     public void build() {
         Stream<String> build = Stream.<String>builder()
-                .add("Mike")
-                .add("Joe")
-                .build();
+            .add("Mike")
+            .add("Joe")
+            .build();
     }
 }
